+++ AppExamples/CleverDeal.React/src/Components/ContentDistribution/index.js
@@ -1,21 +1,3 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from "react-router-dom";
-import ContentDistribution from './App';
-import reportWebVitals from './reportWebVitals';
+import { ContentDistribution } from "./ContentDistribution";
 
-// const root = ReactDOM.createRoot(document.getElementById('root'));
-const root = createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <ContentDistribution />
-    </BrowserRouter>
-  </React.StrictMode>
-);
-
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+export default ContentDistribution;
